Separate Controller from View in file createEvent.jsx
